Revert "ElmyraService: Remove search button from settings activity"
This reverts commit f6006984a700a3fefb781c07ef05f51e9d29ada5.

As confusing as the settings search hierarchy may be, it's still far
more convenient to show the search button on every screen than to force
the user to go all the way back to the settings home page just to
search.

We've reverted this change on the ROM side, so revert it in
ElmyraService as well to make the settings UI match.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: I44b400256aef374241cec00485d3ce423bb8ce9c