Convert service to singleton for system user
Active Edge is a shared CHRE resource, so we can't have multiple
services controlling it when multiple profiles and users are active.
Convert the service and boot receiver to singletons that only run in the
system user to fix conflicts between profiles.

This isn't ideal, and we still need to remove the Active Edge preference
from secondary users, but it fixes the gesture for now.

TODO: Add support for arbitrating access and settings across users
Change-Id: If3cfd4f60ad5a430137d4bcaa68ba35c4add5693